Changed how filenames are formatted on creation + Fixed grenade positions showing incorrectly in json

+++ TopStatsWaffle/Models/GrenadesSpecificStats.cs
@@ -10,9 +10,9 @@ namespace TopStatsWaffle.Models
     {
         public string NadeType { get; set; }
         public long SteamID { get; set; }
-        public int XPosition { get; set; }
-        public int YPosition { get; set; }
-        public int ZPosition { get; set; }
+        public double XPosition { get; set; }
+        public double YPosition { get; set; }
+        public double ZPosition { get; set; }
         public int? NumPlayersFlashed { get; set; }
 
         public GrenadesSpecificStats() { }
